requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File handling
python-magic==0.4.27
boto3==1.34.14
aiofiles==23.2.1

# Email
fastapi-mail==1.4.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (for external analysis software)
httpx==0.26.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
